+++ src/new-sources/new-app.module.ts
@@ -1,11 +1,13 @@
 import { Global, Module } from "@nestjs/common";
 import { NewAppController } from "./new-app.controller";
 import { MainAppModule } from "./m-main-app/main-app.module";
+import { TariffsModule } from "./tariffs";
 
 @Global()
 @Module({
     imports: [
-        MainAppModule.register()
+        MainAppModule.register(),
+        TariffsModule
     ],
     exports: [
     ],
