+++ test/new-tests/m-main-app/commands/create-biz-partner.e2e.ts
@@ -1,10 +1,10 @@
 import { faker } from "@faker-js/faker";
 import { CommandBus } from "@nestjs/cqrs";
 import { getRepositoryToken } from "@nestjs/typeorm";
-import { BizPartner } from "@src/new/m-biz-partners/models";
-import { IBizPartnerVerificationService, IBizPartnerVerificationServiceSymbol } from "@src/new/m-biz-partners/services/biz-partner-verification.service";
-import { CreateBizPartnerCommand, CreateBizPartnerLocationPayload, CreateBizPartnerPayload } from "@src/new/m-biz-partners/usecases/commands/create-biz-partner/create-biz-partner.command";
-import { CustomBizPartnerVerificationService } from "@src/new/m-main-app/services/custom-biz-partner-verification.service";
+import { BizPartner } from "@src/new-sources/m-biz-partners/models";
+import { IBizPartnerVerificationService, IBizPartnerVerificationServiceSymbol } from "@src/new-sources/m-biz-partners/services/biz-partner-verification.service";
+import { CreateBizPartnerCommand, CreateBizPartnerLocationPayload, CreateBizPartnerPayload } from "@src/new-sources/m-biz-partners/usecases/commands/create-biz-partner/create-biz-partner.command";
+import { CustomBizPartnerVerificationService } from "@src/new-sources/m-main-app/services/custom-biz-partner-verification.service";
 import { app } from "@test/test.setup";
 import { Repository } from "typeorm";
